requirements.txt
telethon==1.33.1
aiogram==3.7.0
aiosqlite==0.20.0
python-dotenv==1.0.0
cryptg==0.4.0
regex==2023.12.25
typing-extensions==4.8.0

pytest==8.2.2
pytest-asyncio==0.23.6
